Moved DRR table panel related files in to their own sub directory
