+++ client/src/panels/panel_declarations/results/drr_table/IndicatorDisplayPanel.yaml
