+++ client/src/panels/panel_declarations/results/drr_table/result_flat_table.scss
