+++ client/src/panels/panel_declarations/results/index.js
@@ -1,5 +1,5 @@
 export { declare_explore_results_panel } from "./result_drilldown/result_drilldown";
-export { declare_results_table_panel } from "./result_flat_table";
+export { declare_results_table_panel } from "./drr_table/result_flat_table";
 export { declare_drr_summary_panel } from "./drr_summary";
 export { declare_gov_drr_panel } from "./gov_drr";
 export { declare_gov_dp_panel } from "./gov_dp";
